refactor(Route): Move route config into route folder
